+++ src/commandRegistration.ts
@@ -6,6 +6,7 @@ import { ScriptElement } from './resources/scriptElements';
 import { buildResourceFileContents } from './utils/resourceFileUtils';
 import { openWithKindling } from './commands/kindlingIntegration';
 import { pasteAsJson } from './commands/jsonPaste';
+import { PerspectiveJsonContentProvider, viewPerspectiveJsonCommand } from './providers/perspectiveJsonViewProvider';
 import { provideCodeActions } from './encodedScriptEditing/codeActions';
 import { CodeType, codeTypeMap } from './utils/codeTypes';
 import * as fs from 'fs';
@@ -33,8 +34,15 @@ function registerCodeTypeCommands(
 export function registerCommands(context: vscode.ExtensionContext, dependencyContainer: DependencyContainer, subscriptionManager: any) {
 	const ignitionFileSystemProvider = dependencyContainer.getFileSystemService().ignitionFileSystemProvider;
 
+	// Register the Perspective JSON content provider
+	const perspectiveJsonProvider = new PerspectiveJsonContentProvider();
+	subscriptionManager.add(
+		vscode.workspace.registerTextDocumentContentProvider('perspective-json-view', perspectiveJsonProvider)
+	);
+
 	subscriptionManager.add(vscode.commands.registerCommand('ignition-flint.open-with-kindling', openWithKindling));
 	subscriptionManager.add(vscode.commands.registerCommand('ignition-flint.paste-as-json', pasteAsJson));
+	subscriptionManager.add(vscode.commands.registerCommand('ignition-flint.view-perspective-json', viewPerspectiveJsonCommand));
 	subscriptionManager.add(vscode.languages.registerCodeActionsProvider('json', { provideCodeActions }));
 
 	registerCodeTypeCommands(context, dependencyContainer.getVirtualFileSystemProvider(), openIgnitionCode);